back-end/model.py
from __future__ import annotations

from typing import Dict
import io
import random

from PIL import Image


def predict_image(image_bytes: bytes) -> Dict[str, float | str]:
	"""Placeholder prediction function.
	
	This function can be replaced with real deep learning inference
	(PyTorch, TensorFlow, ONNX Runtime, etc.). For now, it returns
	a random result to enable the pipeline end-to-end.
	
	Args:
		image_bytes: Raw image bytes
		
	Returns:
		Dict with 'label' (str) and 'confidence' (float) keys
	"""
	# Placeholder: return random fruit labels
	# TODO: Replace with actual model inference
	fake_labels = ["Banana A", "Banana B", "Mango A", "Mango B"]
	label = random.choice(fake_labels)
	confidence = round(random.uniform(0.85, 0.99), 2)
	return {"label": label, "confidence": confidence}
