back-end/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pillow==10.4.0
pymongo==4.9.1
python-dotenv==1.0.1
python-multipart
